More tests + refactoring

+++ CheChatApp/views.py
@@ -6,10 +6,11 @@ from CheChatApp.models import Chat, PhoneBook, ChatUser, Message
 from django.contrib.auth.models import User
 from django.http import JsonResponse
 
-    #tested
+
+    # tested
 def user_listing(request):
     """View with the list of users"""
-    #return render(request, 'users/user_listing.html', {'users': User.objects.all()})
+    # return render(request, 'users/user_listing.html', {'users': User.objects.all()})
     list_username = []
     list_id = []
     for user in User.objects.all():
@@ -22,7 +23,7 @@ def user_listing(request):
     }
     return JsonResponse(response)
 
-    #tested
+    # tested
 def get_user_info(request, user_id):
     """Get user info"""
 
@@ -90,7 +91,7 @@ def logout(request):
     auth_logout(request)
     return redirect('login')
 
-
+   #tested
 def info_chat_by_id(request, chat_id):
     chat = Chat.objects.filter(id=chat_id)
 
@@ -106,21 +107,24 @@ def info_chat_by_id(request, chat_id):
 
     return JsonResponse(response)
 
-
+    #tested
 def get_chat_by_user(request):
     """Get chat of the user"""
 
     chat = Chat.objects.filter(participants__in=str(request.user.id)).order_by('-created')
     response = {}
 
-    print(list(chat))
+    #print(list(chat))
     
     if chat.exists():
         response = {
             'state': 'successful',
             'chat': list(chat.values("id"))
         }
-
+    else:
+        response = {
+            'state': 'no chat found'
+        }
     return JsonResponse(response)
 
 
@@ -134,17 +138,20 @@ def get_last_message(request, chat_id):
             'message': list(Message.objects.filter(chat=chat_id).order_by('-timestamp').values("text", "timestamp"))[0]
         }
 
+    elif not is_participants(chat_id, request.user.id):
+        response = {
+            'state': 'not a participant'
+        }
     else:
         response = {
-            'state': 'chat does not exist'
+            'state': 'chat not found'
         }
 
     return JsonResponse(response)
 
-
+    #tested (testare con titolo)
 def new_chat(request, title=""):
     """Create a new"""
-
     # TODO: controllare se l'utente è amico
     if request.user.is_authenticated:
         chat = Chat.objects.create(title=title)
@@ -244,14 +251,17 @@ def add_contact(request, added_user_id):
 
 
 def delete_contact(request, user_to_delete_id):
-    phonebook = PhoneBook.objects.get(owner=request.user)
-    response = {}
+    try:
+        phonebook = PhoneBook.objects.get(owner=request.user)
+    except PhoneBook.DoesNotExist:
+        phonebook = None
+        response = {'state': 'phonebook not found'}
 
-    if User.objects.filter(id=user_to_delete_id).exists() \
+    if User.objects.filter(id=user_to_delete_id).exists() and phonebook is not None \
             and phonebook.contacts.filter(id=user_to_delete_id).exists():
         response = {'state': 'successful'}
         phonebook.contacts.get(id=user_to_delete_id).delete()
-    elif phonebook.contacts.filter(id=user_to_delete_id).exists():
+    elif phonebook is not None and not phonebook.contacts.filter(id=user_to_delete_id).exists():
         response = {'state': 'user does not exist'}
 
     return JsonResponse(response)
